feat(BackEnd): implementar métodos para Heat map

+++ WebAPI/src/main/java/br/com/crescer/monitorveiculos/entidade/HeatMapModel.java
@@ -18,5 +18,5 @@ import lombok.Setter;
 public class HeatMapModel {
 
     private Camera camera;
-    private Long fator;
+    private Double fator;
 }
